allow JAI to work without nvidia partner package
