+++ packages/jupyter-ai-magics/jupyter_ai_magics/providers.py
@@ -38,7 +38,6 @@ from langchain_community.llms import (
     OpenAI,
     SagemakerEndpoint,
 )
-from langchain_nvidia_ai_endpoints import ChatNVIDIA
 
 # this is necessary because `langchain.pydantic_v1.main` does not include
 # `ModelMetaclass`, as it is not listed in `__all__` by the `pydantic.main`
@@ -859,23 +858,3 @@ class QianfanProvider(BaseProvider, QianfanChatEndpoint):
     model_id_key = "model_name"
     pypi_package_deps = ["qianfan"]
     auth_strategy = MultiEnvAuthStrategy(names=["QIANFAN_AK", "QIANFAN_SK"])
-
-
-class ChatNVIDIAProvider(BaseProvider, ChatNVIDIA):
-    id = "nvidia-chat"
-    name = "NVIDIA"
-    models = [
-        "playground_llama2_70b",
-        "playground_nemotron_steerlm_8b",
-        "playground_mistral_7b",
-        "playground_nv_llama2_rlhf_70b",
-        "playground_llama2_13b",
-        "playground_steerlm_llama_70b",
-        "playground_llama2_code_13b",
-        "playground_yi_34b",
-        "playground_mixtral_8x7b",
-        "playground_neva_22b",
-        "playground_llama2_code_34b",
-    ]
-    model_id_key = "model"
-    auth_strategy = EnvAuthStrategy(name="NVIDIA_API_KEY")
